short link generator 1 method only
