--- src/main/java/com/websummarizer/Web/Summarizer/controller/Shortlink.java
@@ -1,85 +1,0 @@
-package com.websummarizer.Web.Summarizer.controller;
-
-import com.websummarizer.Web.Summarizer.model.history.HistoryReqAto;
-import com.websummarizer.Web.Summarizer.services.UserServiceImpl;
-import jakarta.servlet.http.HttpSession;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.time.LocalDateTime;
-import java.util.Random;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-
-@RestController
-public class Shortlink {
-    @Autowired
-    BitLyController bitLyController;
-
-    @Autowired
-    HistoryController historyController;
-
-    @Autowired
-    UserServiceImpl userService;
-
-    public String Shortlink(String url, String output, HttpSession session){
-
-        String email = (String) session.getAttribute("username");
-
-        long id = 1l;
-        try {
-            id = userService.getFoundId(email);
-        } catch (Exception e){
-
-        }
-
-        // Generate a short code for the given URL
-        String shortlinkCode = codeShort(url);
-        // Create a new history request object with the generated short code
-        HistoryReqAto historyReqAto = new HistoryReqAto(id, output,url, shortlinkCode, LocalDateTime.now());
-        // Add the history request to the database and get the response
-        historyController.addHistory(historyReqAto);
-
-//        String shortBitly = bitLyController.getBitly("http://***.***.*.***:8080/"+shortlinkCode); //need domain to make this function work,  else this is just the pc ip address to test
-
-        return shortlinkCode;
-
-    }
-
-    // Method to generate a short code for a given URL
-    public String codeShort(String urlOg){
-        Random rand = new Random();
-        urlOg = urlOg + rand.hashCode(); //add random avoiding duplicate code
-        int num = Math.abs(urlOg.hashCode());
-        String dic = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-        String out = "";
-        int len = dic.length();
-        while (num>0){
-            out = dic.charAt(num%len)+out;
-            num = num/len;
-        };
-        return out;
-    }
-
-    //unfinished checkpassword function
-    public boolean checkPassword(String password){
-        String pattern = "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$";
-
-        // Compile the pattern
-        Pattern regex = Pattern.compile(pattern);
-
-        // Create a Matcher object
-        Matcher matcher = regex.matcher(password);
-        return matcher.matches();
-    }
-
-//    //check password
-//        if(!shortlink.checkPassword(password)) {
-//        model.addAttribute("isValid", false);
-//        model.addAttribute("html", "<span class=\"bi bi-exclamation-triangle-fill\"></span>");
-//        model.addAttribute("message", "Password must contain at least 8 characters, 1 uppercase letter, 1 lowercase letter, 1 number, and 1 special character");
-//
-//        return "user/register";
-//    }
-}
